requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
pdf2image
PyMuPDF
langchain
faiss-cpu
openai
tqdm
pydantic
python-docx
chardet